Refactored the go code to smaller files
